Fix up test apps on Windows based on updated API after 'window open' changes. Also removed some warnings - builds everything without any warnings (MSVC8 -W4 at least) now.

+++ tests/win32gl/win32gl.cpp
@@ -267,7 +267,7 @@ LRESULT CALLBACK window_proc( HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam
 			// and call keyboard_event with char code.
 			gLParam = lParam;
 			gWParam = wParam;
-            if (nonprintable_key(wParam, &last_char_code)) 
+            if (nonprintable_key((int)wParam, &last_char_code)) 
 			{
                 // this is not a printable key
                 keyboard_event(last_char_code, LLQtWebKit::KE_KEY_DOWN);
